Projrestructuring, bug fixes, terrain gen tweaks
