--- Assets/Scripts/menu/WorldDataTransfer.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public static class WorldDataTransfer
-{
-    public static string worldName;
-    public static int worldSeed;
-    public static string worldDate;
-    //Add more as needed.
-}
