+++ Assets/Scripts/util/Config.cs
@@ -4,5 +4,17 @@ using UnityEngine;
 
 public static class Config
 {
+    //Establish the constants
     public const int chunkSize = 32;
+
+    //Global Variables
+    public static bool isPaused = false;
+}
+
+public static class Gates
+{
+    public static Gate registryGate = Gate.Open;
+    public static Gate biomeClimateRegistryGate = Gate.Open;
+
+
 }
