+++ Assets/Scripts/util/Gate.cs
@@ -1,0 +1,5 @@
+public enum Gate{
+    Closed,
+    Open,
+    Ajar    //probably not necessary but never hurts to have a gray area.
+}
