+++ Assets/Scripts/util/Gate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1c738f7eeffdf24eadb177829a56f98
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
