--- Assets/Scripts/world/Legacy.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 462af4f18ddce0f46ba65a929c996138
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
