--- Assets/Scripts/world/Legacy/MapDisplay.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7ea7347c03428b04381ce27fc152314e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
