--- Assets/Scripts/world/Legacy/MapGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bcb02a8e2bc3fb34fbbb3f386688d1b3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
