+++ Assets/Scripts/world/chunk/ChunkLoader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3edd0d7254f491d4dbddbae1b4402c41
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
