--- Assets/Scripts/world/gen/ChunkManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 52134c9c0e3b14c4d83a3219f8e503ca
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
