--- Assets/Scripts/world/global/Gate.cs
@@ -1,5 +1,0 @@
-public enum Gate{
-    Closed,
-    Open,
-    Ajar    //probably not necessary but never hurts to have a gray area.
-}
