--- Assets/Scripts/world/global/Gates.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public static class Gates
-{
-    public static Gate registryGate = Gate.Open;
-    public static Gate biomeClimateRegistryGate = Gate.Open;
-
-
-}
