--- Assets/Scripts/world/global/Gates.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7432c4211c09c274e94321ea13e9553e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
