+++ Assets/Scripts/world/global/WorldDataTransfer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 17d1e5c62d05d13489f7eaa2916c2293
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
